refactor: update LayerZeroAdapter and related components for improved functionality
- Enabled IR-based code generation in foundry.toml to address stack depth issues.
- Enhanced LayerZeroAdapter to include msgValue checks and improved command encoding for read requests.
- Added missing keeper field in various adapter and queue functions for better execution context.
- Updated tests to reflect the addition of the keeper parameter and removed obsolete minDstGasForCall tests from StargateAdapter.

+++ packages/chain-bridge/src/helpers/LayerZeroOptionsHelper.sol
@@ -82,8 +82,8 @@ library LayerZeroOptionsHelper {
 
         // For lzRead operations, we need to provide a reasonable calldata size estimate
         // If not provided in adapterParams, use a default reasonable size
-        uint128 calldataSize = adapterParams.calldataSize > 0
-            ? adapterParams.calldataSize
+        uint32 calldataSize = adapterParams.calldataSize > 0
+            ? uint32(adapterParams.calldataSize)
             : 1024; // Default to 1KB for read responses
 
         // Add our LzRead option to the existing or new options
@@ -91,7 +91,7 @@ library LayerZeroOptionsHelper {
             OptionsBuilder.addExecutorLzReadOption(
                 options,
                 gasLimit,
-                calldataSize, // ✅ Use proper calldata size
+                calldataSize, // ✅ Use proper calldata size (uint32)
                 adapterParams.msgValue
             );
     }
